refactor(client): split ShopPage and HomePage into modules
ShopPage (1511 lines) -> shop/{types,art,ShopCard,DetailSheet} + a 510-line
container. HomePage (1185 lines) -> home/{helpers,cards,sheets,CustomLobbySheet}
+ a 366-line container. Sub-components moved verbatim (no behaviour change); all
new modules are under the 500-line guideline. Verified by the existing test suite.
