+++ packages/client/src/components/HomePage.tsx
@@ -1,11 +1,15 @@
-import React, { useEffect, useRef, useState, useCallback } from 'react';
+import { useEffect, useRef, useState } from 'react';
 import { motion, AnimatePresence } from 'framer-motion';
 import { useTranslation } from 'react-i18next';
 import { useGame } from '../contexts/GameContext';
-import { LoginPanel } from './LoginPanel';
-import type { RoomAvailable } from 'colyseus.js';
-import { colors, gradients, shadows, radii, fonts } from '../theme';
-import { GoldButton, Sheet, Pill, SectionLabel, Divider, TextInput, Select, Diamond } from './ui';
+import { colors, fonts } from '../theme';
+import { Sheet, Diamond } from './ui';
+import { useIsShortViewport } from '../utils/useIsShortViewport';
+import { FloatingCard, MiniCard, MiniCardFan, ModeCard } from './home/cards';
+import { SearchingOverlay, RankedConfigSheet, SinglePlayerSheet } from './home/sheets';
+import { CustomLobbySheet } from './home/CustomLobbySheet';
+import { FLOATING_CARDS } from './home/helpers';
+import type { HomeState, SheetKey, RankedOptions } from './home/helpers';
 
 interface HomePageProps {
   discordId?: string;
@@ -14,845 +18,6 @@ interface HomePageProps {
   avatarUrl?: string;
   error?: string | null;
 }
-
-type SheetKey = 'ranked' | 'custom' | 'solo' | null;
-
-interface RankedOptions {
-  mode: 'classic' | 'teams';
-  maxPlayers: number;
-  handSize: number;
-}
-type HomeState = 'home' | 'searching';
-
-// ── helpers ───────────────────────────────────────────────────────────────────
-
-function getAvailablePlayers(mode: string) {
-  return mode === 'teams' ? [4, 6] : [2, 3, 4, 5, 6];
-}
-function getAvailableHandSizes(mode: string, players: number) {
-  if (mode === 'teams') return players === 4 ? [7] : [5];
-  return players <= 4 ? [5, 7] : [5];
-}
-
-// ── Floating intro cards ──────────────────────────────────────────────────────
-
-function FloatingCard({
-  suit,
-  rank,
-  delay,
-  x,
-  rotate,
-  scale,
-}: {
-  suit: string;
-  rank: string;
-  delay: number;
-  x: number;
-  rotate: number;
-  scale: number;
-}) {
-  const isRed = suit === '♥' || suit === '♦';
-  return (
-    <motion.div
-      initial={{ y: -180, opacity: 0, rotate: rotate - 25 }}
-      animate={{ y: 0, opacity: 1, rotate }}
-      transition={{ delay, duration: 0.9, type: 'spring', stiffness: 70, damping: 13 }}
-      style={{
-        position: 'absolute',
-        left: `${x}%`,
-        top: '5%',
-        scale,
-        zIndex: 1,
-        pointerEvents: 'none',
-        filter: 'drop-shadow(0 12px 24px rgba(0,0,0,0.55))',
-      }}
-    >
-      <div
-        style={{
-          width: 64,
-          height: 92,
-          background: gradients.card,
-          borderRadius: 8,
-          border: `1px solid ${colors.gold[700]}`,
-          padding: '6px 7px',
-          color: isRed ? '#a01818' : '#1b150a',
-          fontFamily: fonts.display,
-          display: 'flex',
-          flexDirection: 'column',
-          justifyContent: 'space-between',
-          boxShadow: 'inset 0 0 0 1px rgba(255,255,255,0.5), 0 6px 16px rgba(0,0,0,0.45)',
-        }}
-      >
-        <div style={{ fontSize: 13, fontWeight: 800, lineHeight: 1 }}>
-          <div>{rank}</div>
-          <div style={{ fontSize: 14, marginTop: 1 }}>{suit}</div>
-        </div>
-        <div style={{ fontSize: 22, textAlign: 'center', opacity: 0.22, fontWeight: 900 }}>
-          {suit}
-        </div>
-        <div
-          style={{
-            fontSize: 13,
-            fontWeight: 800,
-            lineHeight: 1,
-            transform: 'rotate(180deg)',
-            alignSelf: 'flex-end',
-          }}
-        >
-          <div>{rank}</div>
-          <div style={{ fontSize: 14, marginTop: 1 }}>{suit}</div>
-        </div>
-      </div>
-    </motion.div>
-  );
-}
-
-const FLOATING_CARDS = [
-  { suit: '♠', rank: 'A', delay: 0.05, x: 4, rotate: -22, scale: 0.85 },
-  { suit: '♥', rank: 'K', delay: 0.15, x: 14, rotate: -10, scale: 0.95 },
-  { suit: '♦', rank: 'Q', delay: 0.1, x: 24, rotate: 4, scale: 0.9 },
-  { suit: '♣', rank: 'J', delay: 0.2, x: 70, rotate: -5, scale: 0.9 },
-  { suit: '♥', rank: '7', delay: 0.12, x: 80, rotate: 12, scale: 0.95 },
-  { suit: '♠', rank: 'A', delay: 0.07, x: 89, rotate: 24, scale: 0.85 },
-];
-
-// ── Searching overlay ─────────────────────────────────────────────────────────
-
-function SearchingOverlay({ onCancel, label }: { onCancel: () => void; label: string }) {
-  const { t } = useTranslation('home');
-  const [elapsed, setElapsed] = useState(0);
-  useEffect(() => {
-    const interval = setInterval(() => setElapsed((e) => e + 1), 1000);
-    return () => clearInterval(interval);
-  }, []);
-  const fmt = (s: number) =>
-    `${String(Math.floor(s / 60)).padStart(2, '0')}:${String(s % 60).padStart(2, '0')}`;
-
-  return (
-    <motion.div
-      initial={{ opacity: 0, scale: 0.9 }}
-      animate={{ opacity: 1, scale: 1 }}
-      exit={{ opacity: 0, scale: 0.9 }}
-      transition={{ type: 'spring', stiffness: 200, damping: 22 }}
-      style={{
-        display: 'flex',
-        flexDirection: 'column',
-        alignItems: 'center',
-        gap: 28,
-        zIndex: 10,
-      }}
-    >
-      {/* Rotating roulette-style ring */}
-      <div style={{ position: 'relative', width: 168, height: 168 }}>
-        <motion.div
-          animate={{ rotate: 360 }}
-          transition={{ duration: 6, repeat: Infinity, ease: 'linear' }}
-          style={{
-            position: 'absolute',
-            inset: 0,
-            borderRadius: '50%',
-            background: `conic-gradient(${colors.gold[500]} 0deg, transparent 60deg, ${colors.gold[500]} 120deg, transparent 180deg, ${colors.gold[500]} 240deg, transparent 300deg)`,
-            mask: 'radial-gradient(circle, transparent 60%, black 62%, black 100%)',
-            WebkitMask: 'radial-gradient(circle, transparent 60%, black 62%, black 100%)',
-            filter: 'drop-shadow(0 0 16px rgba(212,175,55,0.5))',
-          }}
-        />
-        <div
-          style={{
-            position: 'absolute',
-            inset: 16,
-            borderRadius: '50%',
-            background: gradients.velvet,
-            border: `2px solid ${colors.gold[600]}`,
-            boxShadow: `inset 0 0 24px rgba(0,0,0,0.6), ${shadows.goldGlow}`,
-            display: 'flex',
-            alignItems: 'center',
-            justifyContent: 'center',
-            fontSize: 56,
-          }}
-        >
-          ♠
-        </div>
-      </div>
-
-      <div style={{ textAlign: 'center' }}>
-        <div
-          style={{
-            fontFamily: fonts.display,
-            fontWeight: 700,
-            fontSize: 22,
-            color: colors.gold[300],
-            letterSpacing: 4,
-            textTransform: 'uppercase',
-            textShadow: '0 2px 6px rgba(0,0,0,0.7)',
-          }}
-        >
-          {label}
-        </div>
-        <div
-          style={{
-            fontSize: 14,
-            color: colors.ivory[300],
-            marginTop: 8,
-            fontVariantNumeric: 'tabular-nums',
-            letterSpacing: 1,
-          }}
-        >
-          {fmt(elapsed)}
-        </div>
-      </div>
-
-      <GoldButton variant="burgundy" size="md" onClick={onCancel}>
-        {t('searching.cancel')}
-      </GoldButton>
-    </motion.div>
-  );
-}
-
-// ── Ranked config sheet ───────────────────────────────────────────────────────
-
-function RankedConfigSheet({ onFindMatch }: { onFindMatch: (opts: RankedOptions) => void }) {
-  const { t } = useTranslation('home');
-  const [maxPlayers, setMaxPlayers] = useState(4);
-  const [handSize, setHandSize] = useState(5);
-
-  const playerOptions = [2, 3, 4, 5, 6];
-  const handSizeOptions = getAvailableHandSizes('classic', maxPlayers);
-  const effectiveHandSize = handSizeOptions.includes(handSize) ? handSize : handSizeOptions[0]!;
-
-  const handlePlayerChange = (n: number) => {
-    setMaxPlayers(n);
-    const hs = getAvailableHandSizes('classic', n);
-    if (!hs.includes(handSize)) setHandSize(hs[0]!);
-  };
-
-  return (
-    <div style={{ display: 'flex', flexDirection: 'column', gap: 22 }}>
-      <div>
-        <SectionLabel>{t('ranked.playersAtTable')}</SectionLabel>
-        <div style={{ display: 'flex', gap: 8, flexWrap: 'wrap' }}>
-          {playerOptions.map((n) => (
-            <Pill key={n} active={maxPlayers === n} onClick={() => handlePlayerChange(n)}>
-              {n}
-            </Pill>
-          ))}
-        </div>
-        <div
-          style={{
-            fontSize: 11,
-            color: colors.ivory[300],
-            marginTop: 8,
-            opacity: 0.7,
-            fontStyle: 'italic',
-          }}
-        >
-          {t('ranked.gameBegins', { count: maxPlayers })}
-        </div>
-      </div>
-
-      {handSizeOptions.length > 1 && (
-        <div>
-          <SectionLabel>{t('ranked.cardsInHand')}</SectionLabel>
-          <div style={{ display: 'flex', gap: 10 }}>
-            {handSizeOptions.map((s) => (
-              <button
-                key={s}
-                onClick={() => setHandSize(s)}
-                style={{
-                  flex: 1,
-                  padding: '14px 8px',
-                  borderRadius: radii.md,
-                  cursor: 'pointer',
-                  background: effectiveHandSize === s ? gradients.gold : 'rgba(255,255,255,0.03)',
-                  color: effectiveHandSize === s ? colors.ink[900] : colors.ivory[200],
-                  border: `1.5px solid ${
-                    effectiveHandSize === s ? 'rgba(212,175,55,0.85)' : 'rgba(212,175,55,0.18)'
-                  }`,
-                  fontFamily: fonts.display,
-                  fontWeight: 700,
-                  fontSize: 22,
-                  boxShadow:
-                    effectiveHandSize === s
-                      ? `${shadows.engrave}, ${shadows.goldGlow}`
-                      : shadows.engrave,
-                  transition: 'all 0.18s',
-                }}
-              >
-                {s}
-                <div
-                  style={{
-                    fontFamily: fonts.body,
-                    fontSize: 10,
-                    fontWeight: 600,
-                    marginTop: 2,
-                    opacity: 0.7,
-                    letterSpacing: 1.5,
-                    textTransform: 'uppercase',
-                  }}
-                >
-                  {t('ranked.cards')}
-                </div>
-              </button>
-            ))}
-          </div>
-        </div>
-      )}
-
-      <Divider label={t('ranked.stakeRep')} />
-
-      <GoldButton
-        size="lg"
-        block
-        onClick={() => onFindMatch({ mode: 'classic', maxPlayers, handSize: effectiveHandSize })}
-      >
-        {t('ranked.findMatch')}
-      </GoldButton>
-    </div>
-  );
-}
-
-// ── Single Player sheet ───────────────────────────────────────────────────────
-
-const DIFFICULTIES = [
-  {
-    key: 'easy',
-    label: 'Apprentice',
-    icon: 'badge_apprentice.png',
-    desc: 'Plays loosely, makes mistakes.',
-    accent: '#3f8a4a',
-  },
-  {
-    key: 'medium',
-    label: 'Gentleman',
-    icon: 'badge_gentleman.png',
-    desc: 'Plays smart, occasional blunders.',
-    accent: colors.gold[500],
-  },
-  {
-    key: 'hard',
-    label: 'High Roller',
-    icon: 'badge_high_roller.png',
-    desc: 'Ruthless. Calculates every card.',
-    accent: '#a01818',
-  },
-] as const;
-
-function SinglePlayerSheet() {
-  const { createGame } = useGame();
-  const { t } = useTranslation('home');
-  const [loading, setLoading] = useState<string | null>(null);
-
-  const handlePick = async (difficulty: string) => {
-    setLoading(difficulty);
-    await createGame({
-      maxPlayers: 2,
-      isPrivate: true,
-      mode: 'classic',
-      teamSelection: 'random',
-      handSize: 6,
-      botDifficulty: difficulty,
-    });
-    setLoading(null);
-  };
-
-  return (
-    <div style={{ display: 'flex', flexDirection: 'column', gap: 12 }}>
-      <SectionLabel>{t('solo.chooseOpponent')}</SectionLabel>
-      {DIFFICULTIES.map((d) => (
-        <button
-          key={d.key}
-          onClick={() => handlePick(d.key)}
-          disabled={loading !== null}
-          style={{
-            display: 'flex',
-            alignItems: 'center',
-            gap: 16,
-            padding: '16px 18px',
-            borderRadius: radii.md,
-            cursor: loading ? 'default' : 'pointer',
-            background: gradients.panel,
-            border: `1.5px solid ${d.accent}88`,
-            color: colors.ivory[100],
-            textAlign: 'left',
-            transition: 'all 0.15s',
-            boxShadow: `${shadows.low}, ${shadows.engrave}, 0 0 0 1px ${d.accent}33 inset`,
-            opacity: loading && loading !== d.key ? 0.45 : 1,
-          }}
-        >
-          <span
-            style={{
-              width: 48,
-              height: 48,
-              borderRadius: '50%',
-              overflow: 'hidden',
-              display: 'flex',
-              alignItems: 'center',
-              justifyContent: 'center',
-              boxShadow: `0 4px 10px rgba(0,0,0,0.6), 0 0 12px ${d.accent}44`,
-              border: `1.5px solid ${d.accent}`,
-              background: 'rgba(0,0,0,0.3)',
-              flexShrink: 0,
-            }}
-          >
-            {loading === d.key ? (
-              <span style={{ fontSize: 24 }}>⏳</span>
-            ) : (
-              <img
-                src={`/assets/badges/${d.icon}`}
-                alt=""
-                style={{ width: '100%', height: '100%', objectFit: 'cover', display: 'block' }}
-              />
-            )}
-          </span>
-          <div style={{ flex: 1 }}>
-            <div
-              style={{
-                fontFamily: fonts.display,
-                fontWeight: 700,
-                fontSize: 17,
-                letterSpacing: 1,
-                color: colors.gold[300],
-              }}
-            >
-              {t(
-                `solo.${d.key === 'easy' ? 'apprentice' : d.key === 'medium' ? 'gentleman' : 'highRoller'}`,
-              )}
-            </div>
-            <div
-              style={{
-                fontSize: 12,
-                color: colors.ivory[300],
-                marginTop: 4,
-                opacity: 0.85,
-              }}
-            >
-              {t(
-                `solo.${d.key === 'easy' ? 'apprenticeDesc' : d.key === 'medium' ? 'gentlemanDesc' : 'highRollerDesc'}`,
-              )}
-            </div>
-          </div>
-          <span style={{ color: d.accent, fontSize: 22 }}>›</span>
-        </button>
-      ))}
-    </div>
-  );
-}
-
-// ── Custom Lobby sheet ────────────────────────────────────────────────────────
-
-function CustomLobbySheet({
-  discordId,
-  userId,
-  username,
-  avatarUrl,
-}: {
-  discordId?: string;
-  userId?: string;
-  username?: string;
-  avatarUrl?: string;
-}) {
-  const { createGame, joinGame, spectateGame, findPublicGames } = useGame();
-  const { t } = useTranslation('home');
-  const [tab, setTab] = useState<'create' | 'join' | 'browse'>('create');
-  const [isLoading, setIsLoading] = useState(false);
-  const [rooms, setRooms] = useState<RoomAvailable[]>([]);
-  const [joinCode, setJoinCode] = useState('');
-  const [mode, setMode] = useState('classic');
-  const [maxPlayers, setMaxPlayers] = useState(6);
-  const [handSize, setHandSize] = useState(5);
-  const [teamSelection, setTeamSelection] = useState<'random' | 'manual'>('random');
-
-  const fetchRooms = useCallback(async () => {
-    try {
-      setRooms(await findPublicGames());
-    } catch {
-      /* ignore */
-    }
-  }, [findPublicGames]);
-
-  useEffect(() => {
-    // eslint-disable-next-line react-hooks/set-state-in-effect
-    if (tab === 'browse') void fetchRooms();
-  }, [tab, fetchRooms]);
-
-  return (
-    <div style={{ display: 'flex', flexDirection: 'column', gap: 18 }}>
-      <LoginPanel />
-
-      {/* Tab strip — gold underline */}
-      <div
-        style={{
-          display: 'flex',
-          gap: 2,
-          padding: 4,
-          borderRadius: radii.md,
-          background: 'rgba(0,0,0,0.45)',
-          border: '1px solid rgba(212,175,55,0.15)',
-          boxShadow: shadows.engrave,
-        }}
-      >
-        {(['create', 'join', 'browse'] as const).map((tabKey) => {
-          const active = tab === tabKey;
-          return (
-            <button
-              key={tabKey}
-              onClick={() => setTab(tabKey)}
-              style={{
-                flex: 1,
-                padding: '11px 0',
-                borderRadius: radii.sm,
-                border: 'none',
-                cursor: 'pointer',
-                fontFamily: fonts.display,
-                fontWeight: 700,
-                fontSize: 12,
-                letterSpacing: 2,
-                textTransform: 'uppercase',
-                background: active ? gradients.gold : 'transparent',
-                color: active ? colors.ink[900] : colors.ivory[300],
-                textShadow: active ? '0 1px 0 rgba(255,255,255,0.3)' : undefined,
-                boxShadow: active ? shadows.engrave : 'none',
-                transition: 'all 0.18s',
-              }}
-            >
-              {t(`custom.${tabKey}`)}
-            </button>
-          );
-        })}
-      </div>
-
-      {tab === 'create' && (
-        <div style={{ display: 'flex', flexDirection: 'column', gap: 14 }}>
-          <div>
-            <SectionLabel>{t('custom.mode')}</SectionLabel>
-            <Select
-              value={mode}
-              onChange={(e) => {
-                const m = e.target.value;
-                setMode(m);
-                const vp = getAvailablePlayers(m);
-                const np = vp.includes(maxPlayers) ? maxPlayers : vp[0]!;
-                setMaxPlayers(np);
-                const vh = getAvailableHandSizes(m, np);
-                if (!vh.includes(handSize)) setHandSize(vh[0]!);
-              }}
-            >
-              <option value="classic">{t('custom.modeClassic')}</option>
-              <option value="teams">{t('custom.modeTeams')}</option>
-            </Select>
-          </div>
-          <div>
-            <SectionLabel>{t('custom.players')}</SectionLabel>
-            <Select
-              value={maxPlayers}
-              onChange={(e) => {
-                const n = parseInt(e.target.value);
-                setMaxPlayers(n);
-                const vh = getAvailableHandSizes(mode, n);
-                if (!vh.includes(handSize)) setHandSize(vh[0]!);
-              }}
-            >
-              {getAvailablePlayers(mode).map((n) => (
-                <option key={n} value={n}>
-                  {t('custom.nPlayers', { count: n })}
-                </option>
-              ))}
-            </Select>
-          </div>
-          <div>
-            <SectionLabel>{t('custom.handSize')}</SectionLabel>
-            <Select
-              value={handSize}
-              onChange={(e) => setHandSize(parseInt(e.target.value))}
-              disabled={getAvailableHandSizes(mode, maxPlayers).length <= 1}
-            >
-              {getAvailableHandSizes(mode, maxPlayers).map((s) => (
-                <option key={s} value={s}>
-                  {t('custom.nCards', { count: s })}
-                </option>
-              ))}
-            </Select>
-          </div>
-          {mode === 'teams' && (
-            <div>
-              <SectionLabel>{t('custom.teamAssignment')}</SectionLabel>
-              <Select
-                value={teamSelection}
-                onChange={(e) => setTeamSelection(e.target.value as 'random' | 'manual')}
-              >
-                <option value="random">{t('custom.randomTeams')}</option>
-                <option value="manual">{t('custom.manualSelection')}</option>
-              </Select>
-            </div>
-          )}
-          <GoldButton
-            size="lg"
-            block
-            loading={isLoading}
-            onClick={async () => {
-              setIsLoading(true);
-              await createGame({
-                maxPlayers,
-                isPrivate: true,
-                mode,
-                teamSelection,
-                handSize,
-                ...(discordId ? { discordId } : {}),
-                ...(userId ? { userId } : {}),
-                ...(username ? { username } : {}),
-                ...(avatarUrl ? { avatarUrl } : {}),
-              });
-              setIsLoading(false);
-            }}
-          >
-            {t('custom.openTable')}
-          </GoldButton>
-        </div>
-      )}
-
-      {tab === 'join' && (
-        <div style={{ display: 'flex', flexDirection: 'column', gap: 14 }}>
-          <div>
-            <SectionLabel>{t('custom.inviteCode')}</SectionLabel>
-            <TextInput
-              value={joinCode}
-              onChange={(e) => setJoinCode(e.target.value.toUpperCase())}
-              placeholder="ABC123"
-              maxLength={8}
-              style={{
-                letterSpacing: 6,
-                fontSize: 22,
-                fontWeight: 900,
-                fontFamily: fonts.display,
-                textAlign: 'center',
-                color: colors.gold[300],
-              }}
-            />
-          </div>
-          <GoldButton
-            size="lg"
-            block
-            loading={isLoading}
-            disabled={!joinCode.trim()}
-            onClick={async () => {
-              if (!joinCode.trim()) return;
-              setIsLoading(true);
-              await joinGame(joinCode.trim(), discordId, userId, username, avatarUrl);
-              setIsLoading(false);
-            }}
-          >
-            {t('custom.takeSeat')}
-          </GoldButton>
-        </div>
-      )}
-
-      {tab === 'browse' && (
-        <div style={{ display: 'flex', flexDirection: 'column', gap: 10 }}>
-          <div style={{ display: 'flex', justifyContent: 'flex-end' }}>
-            <button
-              onClick={fetchRooms}
-              style={{
-                background: 'none',
-                border: 'none',
-                color: colors.gold[400],
-                fontFamily: fonts.display,
-                fontWeight: 700,
-                fontSize: 11,
-                letterSpacing: 2,
-                textTransform: 'uppercase',
-                cursor: 'pointer',
-              }}
-            >
-              {t('common:actions.refresh')}
-            </button>
-          </div>
-          {rooms.length === 0 ? (
-            <div
-              style={{
-                textAlign: 'center',
-                padding: '40px 0',
-                color: colors.ivory[300],
-                fontSize: 13,
-                opacity: 0.7,
-                fontStyle: 'italic',
-              }}
-            >
-              {t('custom.lounge')}
-              <br />
-              {t('custom.loungeHint')}
-            </div>
-          ) : (
-            rooms.map((r) => {
-              const meta = r.metadata as Record<string, unknown> | null;
-              const pc = (meta?.playerCount as number) ?? r.clients;
-              const mx = (meta?.maxPlayers as number) ?? r.maxClients;
-              const phase = (meta?.phase as string) ?? 'waiting';
-              const full = pc >= mx;
-              return (
-                <div
-                  key={r.roomId}
-                  style={{
-                    display: 'flex',
-                    alignItems: 'center',
-                    justifyContent: 'space-between',
-                    padding: '14px 16px',
-                    borderRadius: radii.md,
-                    background: gradients.panel,
-                    border: '1px solid rgba(212,175,55,0.2)',
-                    boxShadow: shadows.engrave,
-                  }}
-                >
-                  <div>
-                    <div
-                      style={{
-                        fontFamily: fonts.display,
-                        fontWeight: 700,
-                        fontSize: 14,
-                        color: colors.gold[300],
-                        letterSpacing: 1.5,
-                      }}
-                    >
-                      {t('custom.tableLabel', { id: r.roomId.slice(0, 6) })}
-                    </div>
-                    <div
-                      style={{
-                        fontSize: 11,
-                        color: colors.ivory[300],
-                        marginTop: 3,
-                        opacity: 0.75,
-                        letterSpacing: 0.3,
-                      }}
-                    >
-                      {t('custom.seatedOf', {
-                        count: pc,
-                        max: mx,
-                        mode:
-                          meta?.mode === 'teams' ? t('custom.modeTeams') : t('custom.modeClassic'),
-                      })}
-                    </div>
-                  </div>
-                  <div style={{ display: 'flex', gap: 8 }}>
-                    <GoldButton
-                      size="sm"
-                      variant={full ? 'ghost' : 'gold'}
-                      disabled={isLoading || full}
-                      onClick={() => joinGame(r.roomId, discordId, userId, username, avatarUrl)}
-                    >
-                      {t('custom.join')}
-                    </GoldButton>
-                    <GoldButton
-                      size="sm"
-                      variant="burgundy"
-                      disabled={phase !== 'playing'}
-                      onClick={() => spectateGame(r.roomId)}
-                    >
-                      {t('custom.watch')}
-                    </GoldButton>
-                  </div>
-                </div>
-              );
-            })
-          )}
-        </div>
-      )}
-    </div>
-  );
-}
-
-// ── Mode Button (large, on home screen) ───────────────────────────────────────
-
-const ModeCard: React.FC<{
-  icon: string;
-  title: string;
-  subtitle: string;
-  accent: string;
-  delay: number;
-  onClick: () => void;
-}> = ({ icon, title, subtitle, accent, delay, onClick }) => (
-  <motion.button
-    initial={{ opacity: 0, y: 18 }}
-    animate={{ opacity: 1, y: 0 }}
-    transition={{ delay, type: 'spring', stiffness: 160, damping: 20 }}
-    whileHover={{ y: -4 }}
-    whileTap={{ scale: 0.97 }}
-    onClick={onClick}
-    style={{
-      background: gradients.panel,
-      border: `1.5px solid ${accent}`,
-      borderRadius: radii.md,
-      padding: '16px 18px',
-      color: colors.ivory[100],
-      cursor: 'pointer',
-      display: 'flex',
-      alignItems: 'center',
-      gap: 14,
-      flex: '1 1 180px',
-      minWidth: 175,
-      boxShadow: `${shadows.mid}, ${shadows.engrave}`,
-      textAlign: 'left',
-      position: 'relative',
-      overflow: 'hidden',
-    }}
-  >
-    <div
-      style={{
-        position: 'absolute',
-        top: 0,
-        right: 0,
-        width: 90,
-        height: 90,
-        background: `radial-gradient(circle at top right, ${accent}55, transparent 60%)`,
-        pointerEvents: 'none',
-      }}
-    />
-    <span
-      style={{
-        width: 44,
-        height: 44,
-        display: 'flex',
-        alignItems: 'center',
-        justifyContent: 'center',
-        flexShrink: 0,
-        filter: `drop-shadow(0 2px 8px ${accent}aa)`,
-      }}
-    >
-      {icon.endsWith('.png') ? (
-        <img
-          src={`/assets/home/${icon}`}
-          alt=""
-          style={{ width: '100%', height: '100%', objectFit: 'contain' }}
-        />
-      ) : (
-        <span style={{ fontSize: 30 }}>{icon}</span>
-      )}
-    </span>
-    <div>
-      <div
-        style={{
-          fontFamily: fonts.display,
-          fontWeight: 700,
-          fontSize: 15,
-          letterSpacing: 1.5,
-          textTransform: 'uppercase',
-          color: colors.gold[300],
-        }}
-      >
-        {title}
-      </div>
-      <div
-        style={{
-          fontSize: 11,
-          color: colors.ivory[300],
-          marginTop: 3,
-          opacity: 0.8,
-          letterSpacing: 0.3,
-        }}
-      >
-        {subtitle}
-      </div>
-    </div>
-  </motion.button>
-);
 
 // ── Main HomePage ─────────────────────────────────────────────────────────────
 
@@ -865,6 +30,7 @@ export const HomePage: React.FC<HomePageProps> = ({
 }) => {
   const { joinOrCreateGame, leaveGame } = useGame();
   const { t } = useTranslation('home');
+  const short = useIsShortViewport();
   const [phase, setPhase] = useState<'intro' | 'home'>('intro');
   const [homeState, setHomeState] = useState<HomeState>('home');
   const [sheet, setSheet] = useState<SheetKey>(null);
@@ -933,10 +99,18 @@ export const HomePage: React.FC<HomePageProps> = ({
         display: 'flex',
         flexDirection: 'column',
         alignItems: 'center',
-        justifyContent: 'center',
+        // Short (landscape) viewports: let content scroll from the top instead of
+        // centering + clipping it behind the bottom nav.
+        justifyContent: short ? 'flex-start' : 'center',
         position: 'relative',
-        overflow: 'hidden',
-        paddingBottom: 96,
+        overflowX: 'hidden',
+        overflowY: short ? 'auto' : 'hidden',
+        paddingBottom: 'calc(96px + env(safe-area-inset-bottom, 0px))',
+        paddingTop: short
+          ? 'calc(env(safe-area-inset-top, 0px) + 16px)'
+          : 'env(safe-area-inset-top, 0px)',
+        paddingLeft: 'env(safe-area-inset-left, 0px)',
+        paddingRight: 'env(safe-area-inset-right, 0px)',
       }}
     >
       {/* Faint repeating diamond engraving */}
@@ -952,17 +126,21 @@ export const HomePage: React.FC<HomePageProps> = ({
         }}
       />
 
-      {/* Floating intro cards */}
-      {FLOATING_CARDS.map((c, i) => (
-        <FloatingCard key={i} {...c} />
-      ))}
+      {/* Floating intro cards — hidden on short/landscape viewports where they'd
+          overlap the logo and steal vertical space. */}
+      {!short && FLOATING_CARDS.map((c, i) => <FloatingCard key={i} {...c} />)}
 
       {/* Logo */}
       <motion.div
         initial={{ opacity: 0, y: 30 }}
         animate={{ opacity: 1, y: 0 }}
         transition={{ delay: 0.4, duration: 0.8, type: 'spring' }}
-        style={{ textAlign: 'center', zIndex: 10, marginBottom: 52, position: 'relative' }}
+        style={{
+          textAlign: 'center',
+          zIndex: 10,
+          marginBottom: short ? 20 : 52,
+          position: 'relative',
+        }}
       >
         {/* Ornament above */}
         <div
@@ -1013,45 +191,6 @@ export const HomePage: React.FC<HomePageProps> = ({
           DURAK
         </motion.div>
 
-        {/* Ornament below */}
-        <div
-          style={{
-            display: 'flex',
-            alignItems: 'center',
-            justifyContent: 'center',
-            gap: 10,
-            marginTop: 6,
-          }}
-        >
-          <div
-            style={{
-              width: 40,
-              height: 1,
-              background: 'linear-gradient(90deg, transparent, rgba(212,175,55,0.6))',
-            }}
-          />
-          <span
-            style={{
-              fontFamily: fonts.display,
-              fontSize: 11,
-              fontStyle: 'italic',
-              color: colors.ivory[200],
-              letterSpacing: 5,
-              textTransform: 'uppercase',
-              opacity: 0.8,
-            }}
-          >
-            {t('hero.subtitle')}
-          </span>
-          <div
-            style={{
-              width: 40,
-              height: 1,
-              background: 'linear-gradient(90deg, rgba(212,175,55,0.6), transparent)',
-            }}
-          />
-        </div>
-
         {error && (
           <div
             style={{
@@ -1089,7 +228,7 @@ export const HomePage: React.FC<HomePageProps> = ({
               display: 'flex',
               flexDirection: 'column',
               alignItems: 'center',
-              gap: 28,
+              gap: short ? 16 : 28,
             }}
           >
             {/* Hero PLAY chip — looks like a casino chip */}
@@ -1098,8 +237,8 @@ export const HomePage: React.FC<HomePageProps> = ({
               whileTap={{ scale: 0.92 }}
               onClick={() => setSheet('ranked')}
               style={{
-                width: 188,
-                height: 188,
+                width: short ? 132 : 188,
+                height: short ? 132 : 188,
                 borderRadius: '50%',
                 background:
                   'repeating-conic-gradient(from 0deg, #d4af37 0deg 30deg, #8b6914 30deg 60deg)',
@@ -1157,26 +296,26 @@ export const HomePage: React.FC<HomePageProps> = ({
               }}
             >
               <ModeCard
-                icon="home_ranked.png"
+                icon={<MiniCard suit="♠" rank="A" size={32} />}
                 title={t('modes.ranked')}
                 subtitle={t('modes.rankedSub')}
-                accent="rgba(212,175,55,0.55)"
+                accent="rgba(212,175,55,0.45)"
                 delay={0.1}
                 onClick={() => setSheet('ranked')}
               />
               <ModeCard
-                icon="home_custom.png"
+                icon={<MiniCardFan size={26} />}
                 title={t('modes.custom')}
                 subtitle={t('modes.customSub')}
-                accent="rgba(139,33,33,0.6)"
+                accent="rgba(212,175,55,0.45)"
                 delay={0.17}
                 onClick={() => setSheet('custom')}
               />
               <ModeCard
-                icon="home_solo.png"
+                icon={<MiniCard suit="♥" rank="A" size={32} />}
                 title={t('modes.solo')}
                 subtitle={t('modes.soloSub')}
-                accent="rgba(63,138,74,0.55)"
+                accent="rgba(212,175,55,0.45)"
                 delay={0.24}
                 onClick={() => setSheet('solo')}
               />
@@ -1189,12 +328,7 @@ export const HomePage: React.FC<HomePageProps> = ({
         {sheet === 'ranked' && (
           <Sheet
             title={t('ranked.title')}
-            icon={
-              <img
-                src="/assets/home/home_ranked.png"
-                style={{ width: 26, height: 26, objectFit: 'contain' }}
-              />
-            }
+            icon={<MiniCard suit="♠" rank="A" size={20} />}
             onClose={() => setSheet(null)}
           >
             <RankedConfigSheet onFindMatch={handleFindMatch} />
@@ -1203,12 +337,7 @@ export const HomePage: React.FC<HomePageProps> = ({
         {sheet === 'custom' && (
           <Sheet
             title={t('custom.title')}
-            icon={
-              <img
-                src="/assets/home/home_custom.png"
-                style={{ width: 26, height: 26, objectFit: 'contain' }}
-              />
-            }
+            icon={<MiniCardFan size={16} />}
             onClose={() => setSheet(null)}
           >
             <CustomLobbySheet
@@ -1222,12 +351,7 @@ export const HomePage: React.FC<HomePageProps> = ({
         {sheet === 'solo' && (
           <Sheet
             title={t('solo.title')}
-            icon={
-              <img
-                src="/assets/home/home_solo.png"
-                style={{ width: 26, height: 26, objectFit: 'contain' }}
-              />
-            }
+            icon={<MiniCard suit="♥" rank="A" size={20} />}
             onClose={() => setSheet(null)}
           >
             <SinglePlayerSheet />
